requirements.txt
Django
Mysql-python
python-memcached
pillow
django-render-json
django-tables3
underscore.py
